+++ assets/js/modules/analytics-4/components/setup/SetupLayout/index.tsx
@@ -1,0 +1,45 @@
+/**
+ * Analytics setup layout component.
+ *
+ * Site Kit by Google, Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Internal dependencies
+ */
+import {
+	DefaultModuleSetup,
+	type ModuleSetupLayoutProps,
+} from '@/js/components/setup';
+import { useFeature } from '@/js/hooks/useFeature';
+import useQueryArg from '@/js/hooks/useQueryArg';
+import { MODULE_SLUG_ANALYTICS_4 } from '@/js/modules/analytics-4/constants';
+import InitialSetupLayout from './InitialSetupLayout';
+
+export default function SetupLayout( { moduleSlug }: ModuleSetupLayoutProps ) {
+	const setupFlowRefreshEnabled = useFeature( 'setupFlowRefresh' );
+	const [ showProgress ] = useQueryArg( 'showProgress' );
+
+	const isInitialSetupFlow =
+		setupFlowRefreshEnabled &&
+		moduleSlug === MODULE_SLUG_ANALYTICS_4 &&
+		showProgress === 'true';
+
+	if ( ! isInitialSetupFlow ) {
+		return <DefaultModuleSetup moduleSlug={ moduleSlug } />;
+	}
+
+	return <InitialSetupLayout moduleSlug={ moduleSlug } />;
+}
